Created new tileset + scene
There is a new scene called 'TestScene' located in Assets/Test/. This scene is a duplicate of scene2 with the addition of a tileset prototype. It currently generates and removes tiles infinitely from a selection of 3 presets.
The directory also contains scripts and assets used to generate the tiles. The files will be moved to the correct directories later.